+++ src/main/java/courgette/runtime/CourgetteReporter.java
@@ -1,136 +1,180 @@
 package courgette.runtime;
 
+import courgette.runtime.report.JsonReportParser;
+import courgette.runtime.utils.FileUtils;
 import io.cucumber.messages.types.Envelope;
 
 import java.io.File;
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
+import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.stream.Collectors;
 
-import static courgette.runtime.CourgetteException.printExceptionStackTrace;
 import static courgette.runtime.utils.FileUtils.writeFile;
 
 class CourgetteReporter {
-
-    private final Map<String, CopyOnWriteArrayList<String>> reports;
-    private final List<Envelope> messages;
-    private final CourgetteRuntimeOptions courgetteRuntimeOptions;
     private final CourgetteProperties courgetteProperties;
+    private final List<CourgetteReportOptions> courgetteReportOptions;
+    private final Map<String, Collection<String>> errors = new HashMap<>();
+    private JsonReportParser jsonReportParser;
 
-    CourgetteReporter(Map<String, CopyOnWriteArrayList<String>> reports,
-                      Map<io.cucumber.core.gherkin.Feature, List<List<Envelope>>> reportMessages,
-                      CourgetteRuntimeOptions courgetteRuntimeOptions,
-                      CourgetteProperties courgetteProperties) {
-
-        this.reports = reports;
-        this.courgetteRuntimeOptions = courgetteRuntimeOptions;
+    public CourgetteReporter(List<CourgetteReportOptions> courgetteReportOptions, CourgetteProperties courgetteProperties) {
         this.courgetteProperties = courgetteProperties;
-
-        this.messages = createMessages(reportMessages);
-
-        if (hasMessages()) {
-            createNdJsonReport(this.messages);
-        }
+        this.courgetteReportOptions = courgetteReportOptions;
     }
 
-    void createCucumberReport(String reportFile, boolean mergeTestCaseName) {
+    Optional<String> createCucumberReports(List<String> reportFiles, boolean publishReport) {
+        final Optional<String> htmlReportFile = reportFiles.stream().filter(report -> report.endsWith(".html")).findFirst();
+        final Optional<String> jsonReportFile = reportFiles.stream().filter(report -> report.endsWith(".json")).findFirst();
+        final Optional<String> ndJsonReportFile = reportFiles.stream().filter(report -> report.endsWith(".ndjson")).findFirst();
+        final Optional<String> xmlReportFile = reportFiles.stream().filter(report -> report.endsWith(".xml")).findFirst();
 
-        if (reportFile != null && !reports.isEmpty()) {
+        String jsonFile = jsonReportFile.orElseGet(() -> FileUtils.createTempFile("json").getPath());
 
-            final List<String> reportData = getReportData();
+        jsonReportParser = new JsonReportParser(jsonFile, courgetteProperties.isFeatureRunLevel());
 
-            final boolean isHtml = reportFile.endsWith(".html");
-            final boolean isJson = reportFile.endsWith(".json");
-            final boolean isNdJson = reportFile.endsWith(".ndjson");
-            final boolean isXml = reportFile.endsWith(".xml");
+        List<String> jsonReports = courgetteReportOptions.stream()
+                .map(CourgetteReportOptions::getJsonFile)
+                .collect(Collectors.toList());
 
-            if (isHtml && courgetteProperties.isCucumberHtmlReportEnabled() && hasMessages()) {
-                CucumberHtmlReporter.createReport(reportFile, messages);
-            }
+        CucumberJsonReporter cucumberJsonReporter = new CucumberJsonReporter(jsonFile, jsonReports.size());
+        jsonReports.forEach(cucumberJsonReporter::readAndWriteReport);
 
-            if (isJson) {
-                reportData.removeIf(report -> !report.startsWith("["));
-                CucumberJsonReporter.createReport(reportFile, reportData);
-            }
+        if (cucumberJsonReporter.hasErrors()) {
+            errors.put(jsonFile, cucumberJsonReporter.getErrors().values());
+        }
 
-            if (isNdJson && hasMessages()) {
-                if (!reportFile.equals(courgetteRuntimeOptions.getCourgetteReportNdJson())) {
-                    CucumberNdJsonReporter.copyReport(courgetteRuntimeOptions.getCourgetteReportNdJson(), reportFile);
-                }
-            }
+        Map<String, List<CourgetteReportOptions>> groupedReports = courgetteReportOptions.stream()
+                .collect(Collectors.groupingBy(CourgetteReportOptions::getFeatureId));
 
-            if (isXml) {
-                reportData.removeIf(report -> !report.startsWith("<?xml"));
-                CucumberXmlReporter.createReport(reportFile, reportData, mergeTestCaseName, courgetteProperties.isReportPortalPluginEnabled());
+        String ndJsonFile = ndJsonReportFile.orElseGet(() -> FileUtils.createTempFile("ndjson").getPath());
+
+        CucumberNdJsonReporter cucumberNdJsonReporter = new CucumberNdJsonReporter(ndJsonFile, groupedReports.size());
+
+        CucumberHtmlReporter cucumberHtmlReporter;
+        if (htmlReportFile.isPresent() && courgetteProperties.isCucumberHtmlReportEnabled()) {
+            cucumberHtmlReporter = new CucumberHtmlReporter(htmlReportFile.get(), groupedReports.size());
+        } else {
+            cucumberHtmlReporter = null;
+        }
+
+        groupedReports.entrySet()
+                .stream()
+                .sorted(Map.Entry.comparingByKey())
+                .forEach((entry) -> {
+                    Map<String, List<List<Envelope>>> messages = cucumberNdJsonReporter.readReports(entry);
+
+                    if (!messages.isEmpty()) {
+                        List<Envelope> runLevelMessages = createRunLevelMessages(messages);
+
+                        cucumberNdJsonReporter.writeReport(runLevelMessages);
+
+                        if (cucumberHtmlReporter != null) {
+                            cucumberHtmlReporter.writeReport(runLevelMessages);
+                        }
+                    }
+                });
+
+        if (cucumberNdJsonReporter.hasErrors()) {
+            errors.put(ndJsonFile, cucumberNdJsonReporter.getErrors().values());
+        }
+
+        if (cucumberHtmlReporter != null && cucumberHtmlReporter.hasErrors()) {
+            errors.put(htmlReportFile.get(), cucumberHtmlReporter.getErrors().values());
+        }
+
+        if (xmlReportFile.isPresent()) {
+            List<String> xmlReports = courgetteReportOptions.stream()
+                    .map(CourgetteReportOptions::getXmlFile)
+                    .map(report -> report.orElse(null))
+                    .filter(Objects::nonNull)
+                    .collect(Collectors.toList());
+
+            boolean mergeTestCaseName = courgetteProperties.isReportPortalPluginEnabled();
+
+            CucumberXmlReporter cucumberXmlReporter = new CucumberXmlReporter(xmlReportFile.get(), xmlReports.size());
+            xmlReports.forEach(report -> cucumberXmlReporter.readAndWriteReport(report, mergeTestCaseName, courgetteProperties.isReportPortalPluginEnabled()));
+
+            if (cucumberXmlReporter.hasErrors()) {
+                errors.put(xmlReportFile.get(), cucumberXmlReporter.getErrors().values());
             }
         }
+
+        Optional<String> publishedReport = Optional.empty();
+
+        if (publishReport) {
+            publishedReport = publishCucumberReport(ndJsonFile);
+        }
+
+        return publishedReport;
     }
 
-    Optional<String> publishCucumberReport() {
+    public boolean hasErrors() {
+        return !errors.isEmpty();
+    }
+
+    public void deleteTemporaryReports() {
+        courgetteReportOptions.stream().map(CourgetteReportOptions::getJsonFile).forEach(FileUtils::deleteFileSilently);
+        courgetteReportOptions.stream().map(CourgetteReportOptions::getNdJsonFile).forEach(FileUtils::deleteFileSilently);
+        courgetteReportOptions.stream().map(CourgetteReportOptions::getXmlFile).forEach(report -> report.ifPresent(FileUtils::deleteFileSilently));
+        courgetteReportOptions.stream().map(CourgetteReportOptions::getRerunFile).forEach(rerun -> rerun.ifPresent(FileUtils::deleteFileSilently));
+    }
+
+    public JsonReportParser jsonReportParser() {
+        return jsonReportParser;
+    }
+
+    public void createErrorReport() {
+        StringBuilder errors = new StringBuilder();
+        this.errors.forEach((key, value1) -> {
+            errors.append(String.format("Report file: %s\n", key));
+            errors.append("Processing errors:\n");
+            value1.forEach(value -> {
+                errors.append(String.format("\t%s\n", value));
+            });
+            errors.append("\n");
+        });
+
+        String errorReportFile = courgetteProperties.getCourgetteOptions().reportTargetDir() + File.separator + "courgette-report-processing-errors.txt";
+        FileUtils.writeFile(errorReportFile, errors.toString());
+    }
+
+    private Optional<String> publishCucumberReport(String messagesFile) {
         Optional<String> reportUrl = Optional.empty();
 
-        if (courgetteProperties.isCucumberReportPublisherEnabled()
-                && courgetteProperties.isCucumberHtmlReportEnabled()
-                && hasMessages()) {
+        final File ndJsonReport = new File(messagesFile);
 
-            final File ndJsonReport = new File(courgetteRuntimeOptions.getCourgetteReportNdJson());
+        if (ndJsonReport.exists()) {
 
-            if (ndJsonReport.exists()) {
+            CucumberReportPublisher reportPublisher = new CucumberReportPublisher(ndJsonReport);
+            reportUrl = reportPublisher.publish();
 
-                CucumberReportPublisher reportPublisher = new CucumberReportPublisher(ndJsonReport);
-                reportUrl = reportPublisher.publish();
+            StringBuilder out = new StringBuilder();
 
-                StringBuilder out = new StringBuilder();
+            if (reportUrl.isPresent()) {
+                out.append("\n─────────────────────────────────────────────────────────────────────────\n");
+                out.append("Report published at: ").append(Instant.now()).append("\n");
+                out.append("\nCourgette published your Cucumber Report to:\n");
+                out.append(reportUrl.get());
+                out.append("\n─────────────────────────────────────────────────────────────────────────\n");
+                System.out.println(out);
 
-                if (reportUrl.isPresent()) {
-                    out.append("\n─────────────────────────────────────────────────────────────────────────\n");
-                    out.append("Report published at: ").append(Instant.now()).append("\n");
-                    out.append("\nCourgette published your Cucumber Report to:\n");
-                    out.append(reportUrl.get());
-                    out.append("\n─────────────────────────────────────────────────────────────────────────\n");
-                    System.out.println(out);
-
-                    String reportLinkFilename = courgetteProperties.getCourgetteOptions().reportTargetDir() + File.separator + "cucumber-report-link.txt";
-                    writeFile(reportLinkFilename, out.toString());
-                }
+                String reportLinkFilename = courgetteProperties.getCourgetteOptions().reportTargetDir() + File.separator + "cucumber-report-link.txt";
+                writeFile(reportLinkFilename, out.toString());
             }
         }
         return reportUrl;
     }
 
-    private List<String> getReportData() {
-        final List<String> reportData = new ArrayList<>();
+    private List<Envelope> createRunLevelMessages(Map<String, List<List<Envelope>>> messages) {
+        final CourgetteNdJsonCreator ndJsonCreator = new CourgetteNdJsonCreator(messages);
 
-        this.reports.values().removeIf(t -> t.contains(null) || t.contains("null") || t.contains("[]") || t.contains(""));
-
-        final Map<String, CopyOnWriteArrayList<String>> reportMap = new LinkedHashMap<>(this.reports);
-
-        reportMap.values().forEach(report -> {
-            try {
-                reportData.add(report.get(0));
-            } catch (Exception e) {
-                printExceptionStackTrace(e);
-            }
-        });
-        return reportData;
-    }
-
-    private List<Envelope> createMessages(Map<io.cucumber.core.gherkin.Feature, List<List<Envelope>>> reportMessages) {
-        final CourgetteNdJsonCreator ndJsonCreator = new CourgetteNdJsonCreator(reportMessages);
-
-        return courgetteProperties.isFeatureRunLevel() ?
-                ndJsonCreator.createFeatureMessages() : ndJsonCreator.createScenarioMessages();
-    }
-
-    private void createNdJsonReport(List<Envelope> messages) {
-        CucumberNdJsonReporter.createReport(courgetteRuntimeOptions.getCourgetteReportNdJson(), messages);
-    }
-
-    private boolean hasMessages() {
-        return messages != null;
+        return new ArrayList<>(courgetteProperties.isFeatureRunLevel() ?
+                ndJsonCreator.createFeatureMessages() : ndJsonCreator.createScenarioMessages());
     }
 }