+++ src/main/java/courgette/runtime/CourgetteRuntimeOptions.java
@@ -1,13 +1,11 @@
 package courgette.runtime;
 
 import courgette.api.CucumberOptions;
-import courgette.integration.reportportal.ReportPortalProperties;
 import courgette.runtime.utils.SystemPropertyUtils;
 import io.cucumber.core.gherkin.Feature;
 import io.cucumber.core.options.CommandlineOptionsParser;
 import io.cucumber.core.options.RuntimeOptions;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -26,18 +24,20 @@ import static java.util.Arrays.copyOf;
 
 class CourgetteRuntimeOptions {
     private final CourgetteProperties courgetteProperties;
+    private final CourgetteReportOptions courgetteReportOptions;
     private final Feature feature;
     private final CucumberOptions cucumberOptions;
     private final String reportTargetDir;
     private final List<String> runtimeOptions = new ArrayList<>();
-
-    private String rerunFile;
     private String cucumberResourcePath;
 
     private final int instanceId = Math.abs(UUID.randomUUID().hashCode());
 
-    CourgetteRuntimeOptions(CourgetteProperties courgetteProperties, Feature feature) {
+    CourgetteRuntimeOptions(CourgetteProperties courgetteProperties,
+                            CourgetteReportOptions courgetteReportOptions,
+                            Feature feature) {
         this.courgetteProperties = courgetteProperties;
+        this.courgetteReportOptions = courgetteReportOptions;
         this.feature = feature;
         this.cucumberOptions = courgetteProperties.getCourgetteOptions().cucumberOptions();
         this.cucumberResourcePath = determineResourcePath(feature);
@@ -48,6 +48,7 @@ class CourgetteRuntimeOptions {
 
     CourgetteRuntimeOptions(CourgetteProperties courgetteProperties) {
         this.courgetteProperties = courgetteProperties;
+        this.courgetteReportOptions = null;
         this.cucumberOptions = courgetteProperties.getCourgetteOptions().cucumberOptions();
         this.feature = null;
         this.reportTargetDir = courgetteProperties.getCourgetteOptions().reportTargetDir();
@@ -64,16 +65,11 @@ class CourgetteRuntimeOptions {
     }
 
     public String getRerunFile() {
-        return rerunFile;
+        return courgetteReportOptions.getRerunFile().orElse(null);
     }
 
     public String getCucumberRerunFile() {
-        final String cucumberRerunFile = cucumberRerunPlugin.apply(courgetteProperties);
-
-        if (cucumberRerunFile == null) {
-            return getRerunFile();
-        }
-        return cucumberRerunFile;
+        return cucumberRerunPlugin.apply(courgetteProperties);
     }
 
     public List<String> getReportFiles() {
@@ -88,33 +84,8 @@ class CourgetteRuntimeOptions {
         return reportFiles;
     }
 
-    public String getJsonReportFile() {
-        return getReportFiles().stream()
-                .filter(isJsonReportPlugin)
-                .filter(r -> r.contains(courgetteProperties.getSessionId()))
-                .findFirst()
-                .orElse(null);
-    }
-
-    public String getCourgetteReportDataDirectory() {
-        return reportTargetDir + "/courgette-report/data";
-    }
-
     public String getSessionReportsDirectory() {
         return String.format("%s/session-reports/%s/", reportTargetDir, courgetteProperties.getSessionId());
-    }
-
-    public String getCourgetteReportJson() {
-        return String.format("%s/report.json", getCourgetteReportDataDirectory());
-    }
-
-    public String getCourgetteReportNdJson() {
-        return String.format("%s/report.ndjson", getCourgetteReportDataDirectory());
-    }
-
-    public String getCourgetteReportXmlForReportPortal() {
-        final ReportPortalProperties reportPortalProperties = ReportPortalProperties.getInstance();
-        return String.format("%s/%s.xml", getCourgetteReportDataDirectory(), reportPortalProperties.getLaunchName());
     }
 
     public String getCourgetteRunLogJson() {
@@ -146,7 +117,7 @@ class CourgetteRuntimeOptions {
     private final BiFunction<String, String[], String[]> envCucumberOptionParser = (propertyName, cucumberOptions) -> {
         String cucumberOption = SystemPropertyUtils.fromSystemEnvOrProperty("cucumber." + propertyName);
 
-        if (cucumberOption != null && cucumberOption.trim().length() > 0) {
+        if (cucumberOption != null && !cucumberOption.trim().isEmpty()) {
             final List<String> options = new ArrayList<>();
             Arrays.stream(cucumberOption.split(",")).forEach(t -> options.add(t.trim()));
 
@@ -156,23 +127,11 @@ class CourgetteRuntimeOptions {
         return cucumberOptions;
     };
 
-    private String getMultiThreadRerunFile() {
-        return getTempDirectory() + courgetteProperties.getSessionId() + "_rerun_" + getFeatureId(feature) + ".txt";
-    }
-
-    private String getMultiThreadReportFile() {
-        return getTempDirectory() + courgetteProperties.getSessionId() + "_thread_report_" + getFeatureId(feature);
-    }
-
-    private String getFeatureReportFile() {
+    private String getSessionFeatureReportFile() {
         final String featureName = Arrays.stream(feature.getUri().getPath().split("/")).reduce((x, y) -> y)
                 .get().replace(".feature", "")
                 .toLowerCase();
         return getSessionReportsDirectory() + String.format("%s_%s", featureName, instanceId);
-    }
-
-    private String getFeatureId(Feature feature) {
-        return String.format("%s_%s", feature.hashCode(), instanceId);
     }
 
     private final Function<CourgetteProperties, String> cucumberRerunPlugin = (courgetteProperties) -> {
@@ -188,96 +147,39 @@ class CourgetteRuntimeOptions {
 
     private final Predicate<String> isReportPlugin = (plugin) -> plugin.startsWith("html:") || plugin.startsWith("json:") || plugin.startsWith("junit:") || plugin.startsWith("message:");
 
-    private final Predicate<String> isJsonReportPlugin = (plugin) -> plugin.endsWith(".json");
-
     private String[] parsePlugins(String[] plugins) {
-        HashSet<String> pluginCollection = new HashSet<>();
-
-        plugins = addDefaultPlugins(plugins);
-
-        asList(plugins).forEach(plugin -> {
-            if (isReportPlugin.test(plugin)) {
-                if (feature != null) {
-                    pluginCollection.add(plugin);
-
-                    String extension = plugin.substring(0, plugin.indexOf(":"));
-
-                    if (extension.equalsIgnoreCase("junit")) {
-                        pluginCollection.remove(plugin);
-                        final String reportPath = String.format("junit:%s.xml", getMultiThreadReportFile());
-                        pluginCollection.add(reportPath);
-                    } else if (extension.equalsIgnoreCase("message")) {
-                        final String reportPath = String.format("message:%s.ndjson", getMultiThreadReportFile());
-                        pluginCollection.add(reportPath);
-                    } else {
-                        if (!extension.equals("") && !extension.equals("html")) {
-                            final String reportPath = String.format("%s:%s.%s", extension, getMultiThreadReportFile(), extension);
-                            pluginCollection.add(reportPath);
-                        }
-                    }
-                } else {
-                    pluginCollection.add(plugin);
-                }
-            } else {
-                pluginCollection.add(plugin);
-            }
-        });
-
-        Predicate<HashSet<String>> alreadyAddedRerunPlugin = (addedPlugins) -> addedPlugins.stream().anyMatch(p -> p.startsWith("rerun:"));
-
-        if (!alreadyAddedRerunPlugin.test(pluginCollection)) {
-            if (feature != null) {
-                rerunFile = getMultiThreadRerunFile();
-            } else {
-                final String cucumberRerunFile = cucumberRerunPlugin.apply(courgetteProperties);
-                rerunFile = cucumberRerunFile != null ? cucumberRerunFile : String.format("%s/courgette-rerun.txt", reportTargetDir);
-            }
-            pluginCollection.add("rerun:" + rerunFile);
-        }
-
-        if (feature != null && pluginCollection.stream().noneMatch(plugin -> plugin.startsWith("json:"))) {
-            pluginCollection.add(String.format("json:%s.json", getMultiThreadReportFile()));
-        }
-
-        if (courgetteProperties.isReportPortalPluginEnabled()) {
-            if (pluginCollection.stream().noneMatch(plugin -> plugin.contains(getCourgetteReportXmlForReportPortal()))) {
-                pluginCollection.add("junit:" + getCourgetteReportXmlForReportPortal());
-            }
-        }
-
-        if (feature != null) {
-            final String junitReportPlugin = String.format("junit:%s.xml", getMultiThreadReportFile());
-            if (pluginCollection.stream().noneMatch(plugin -> plugin.equals(junitReportPlugin))) {
-                pluginCollection.add(junitReportPlugin);
-            }
-
-            if (courgetteProperties.shouldPersistCucumberJsonReports()) {
-                final String featureReportFile = getFeatureReportFile();
-                pluginCollection.add(String.format("json:%s.json", featureReportFile));
-                pluginCollection.add(String.format("message:%s.ndjson", featureReportFile));
-            }
-        }
-
-        checkDisabledPlugins(pluginCollection);
-
+        HashSet<String> pluginCollection = new HashSet<>(asList(plugins));
+        removeReportPlugins(pluginCollection);
+        addRuntimeReportPlugins(pluginCollection);
+        addRuntimeRerunPlugin(pluginCollection);
         return copyOf(pluginCollection.toArray(), pluginCollection.size(), String[].class);
     }
 
-    private void checkDisabledPlugins(HashSet<String> plugins) {
-        if (plugins.stream().anyMatch(p -> p.startsWith("html"))) {
-            if (!courgetteProperties.isCucumberHtmlReportEnabled()) {
-                plugins.removeIf(p -> p.startsWith("html") || (
-                        !courgetteProperties.shouldPersistCucumberJsonReports() && p.startsWith("message"))
-                );
+    private void addRuntimeReportPlugins(HashSet<String> plugins) {
+        if (feature != null) {
+            plugins.add("json:" + courgetteReportOptions.getJsonFile());
+            plugins.add("message:" + courgetteReportOptions.getNdJsonFile());
+
+            if (courgetteProperties.shouldPersistCucumberJsonReports()) {
+                String sessionFeatureReportFile = getSessionFeatureReportFile();
+                plugins.add(String.format("json:%s.json", sessionFeatureReportFile));
+                plugins.add(String.format("message:%s.ndjson", sessionFeatureReportFile));
             }
+
+            courgetteReportOptions.getXmlFile().ifPresent(plugin -> plugins.add("junit:" + plugin));
         }
     }
 
-    private String[] addDefaultPlugins(String[] plugins) {
-        plugins = Arrays.copyOf(plugins, plugins.length + 2);
-        plugins[plugins.length - 1] = "json:" + getCourgetteReportJson();
-        plugins[plugins.length - 2] = "message:" + getCourgetteReportNdJson();
-        return plugins;
+    private void addRuntimeRerunPlugin(HashSet<String> plugins) {
+        if (feature != null) {
+            courgetteReportOptions.getRerunFile().ifPresent(plugin -> plugins.add("rerun:" + plugin));
+        }
+    }
+
+    private void removeReportPlugins(HashSet<String> plugins) {
+        if (feature != null) {
+            plugins.removeIf(plugin -> plugin.startsWith("html") || plugin.startsWith("json") || plugin.startsWith("message") || plugin.startsWith("junit"));
+        }
     }
 
     private final BiFunction<String, Object, List<String>> optionParser = (name, options) -> {
@@ -315,16 +217,6 @@ class CourgetteRuntimeOptions {
         }
         return featurePaths;
     };
-
-    private String getTempDirectory() {
-        final String fileSeparator = File.separator;
-        final String tmpDir = System.getProperty("java.io.tmpdir");
-
-        if (!tmpDir.endsWith(fileSeparator)) {
-            return tmpDir + fileSeparator;
-        }
-        return tmpDir;
-    }
 
     private String determineResourcePath(Feature feature) {
         return String.format("%s:%s",