+++ src/main/java/courgette/runtime/RunStatus.java
@@ -2,6 +2,5 @@ package courgette.runtime;
 
 public enum RunStatus {
     OK,
-    ERROR,
-    REPORT_PROCESSING_ERROR
+    ERROR
 }